moved converter control strategies to separate file
